requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Bootstrap-Flask==2.2.0
requests==2.31.0
python-dotenv==1.2.1